backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==1.4.41
pydantic==1.10.13
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
transformers==4.35.2
torch==2.2.0
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
SpeechRecognition==3.10.0
gTTS==2.3.2
python-magic==0.4.27
bcrypt==4.0.1
marshmallow-sqlalchemy==0.28.2 